requirements.txt
Flask==2.2.2
twilio==8.9.1
Werkzeug==2.2.2
python-dotenv==1.0.0
